+++ TMC2160A/TMC2160A_DEV.h
@@ -67,8 +67,8 @@ void TMC2160A_dev_Setup(
     SPI_HandleTypeDef* hspi, 
     u8 ch,
     u16 ioBase,
-    s8 (*ioWrite)(u16 addr, const u8 *pDat, u16 nBytes),
-    s8 (*ioRead)(u16 addr, u8 *pDat, u16 nBytes)
+    s32 (*ioWrite)(u16 addr, const u8 *pDat, u16 nBytes),
+    s32 (*ioRead)(u16 addr, u8 *pDat, u16 nBytes)
 );
 
 #endif
